+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Feature: [Connector] API plan changed to API restrictions
  - Feature: [import] E-mail anonymize and encryption
  - Feature: [Import] Matching status for order import
+ - Feature: [import] Support for fractional days in order import url in days parameter
  - BugFix: Hydrate address for FBA orders, when fields are null (amazon_us)
 
 # v2.0.0